move centrality code
